Add README for llm_core module; refactor OpenAI model config structure
