+++ llm_core/llm_core/models/__init__.py
@@ -1,6 +1,6 @@
 from typing import Union
 
-from llm_core.models.openai_model_config import OpenAIModelConfig
+from llm_core.models.providers.openai_model_config import OpenAIModelConfig
 
 
 ModelConfigType = Union[OpenAIModelConfig]
